fix: split glm-4.7-flash into free and paid models
The real bug behind round-2's "test wallet could pin a paid provider of a
free model": glm-4.7-flash was flagged free: true while carrying a paid
embercloud mapping alongside the free zai one. Fix it in the catalog, not
with a runtime guard.

- Split into two root models: glm-4.7-flash (embercloud only, PAYG, not
  free) and glm-4.7-flash-free (zai only, free, zero-priced).
- Add a @llmgateway/models invariant test: every free model must have only
  zero-priced provider mappings (recursive price walk). Audit confirmed
  glm-4.7-flash was the only violator.
- With the data invariant enforced, revert the test-wallet guard and auto
  filter to the existing model-level isModelTrulyFree check (drop the
  defensive per-mapping recursive walk added in the previous commit).
